+++ src/main/java/com/codeactuator/rocket/domain/App.java
@@ -1,4 +1,4 @@
-package com.codeactuator.domain;
+package com.codeactuator.rocket.domain;
 
 /**
  * Hello world!
