+++ src/main/java/com/codeactuator/rocket/dto/UserDTO.java
@@ -1,4 +1,4 @@
-package com.codeactuator.dto;
+package com.codeactuator.rocket.dto;
 
 
 import javax.validation.constraints.NotEmpty;
