+++ src/test/java/com/codeactuator/rocket/domain/AppTest.java
@@ -1,4 +1,4 @@
-package com.codeactuator.domain;
+package com.codeactuator.rocket.domain;
 
 import static org.junit.Assert.assertTrue;
